[config] remove config check in discretizers.builtin.gui.matplotlib
